Renamed all Veritas mentions to Axiom
