+++ lib/data_mapper/support/axiom/adapter/data_objects.rb
@@ -1,6 +1,6 @@
-require 'veritas-do-adapter'
+require 'axiom-do-adapter'
 
-module Veritas
+module Axiom
   module Adapter
 
     # Reopenend to add functionality that should eventually
@@ -22,16 +22,17 @@ module Veritas
 
       # Wrap the given +relation+ with a gateway
       #
-      # @param [Veritas::Relation] relation
+      # @param [Axiom::Relation] relation
       #   the relation to wrap with a gateway
       #
-      # @return [Veritas::Relation::Gateway]
+      # @return [Axiom::Relation::Gateway]
       #
       # @api private
       def gateway(relation)
-        Veritas::Relation::Gateway.new(self, relation)
+        Axiom::Relation::Gateway.new(self, relation)
       end
 
     end # class DataObjects
   end # module Adapter
-end # module Veritas
+end # module Axiom
+
